backend/requirements.txt
# ─── Web Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-multipart==0.0.9

# ─── Database ───────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.15.0
alembic==1.13.1

# ─── Cache / Broker ────────────────────────────────────────
redis==5.0.3
celery[redis]==5.3.6

# ─── Auth ───────────────────────────────────────────────────
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ─── Validation ─────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── Geospatial ─────────────────────────────────────────────
rasterio==1.3.10
geopandas==0.14.3
shapely==2.0.4
pyproj==3.6.1
geojson==3.1.0

# ─── Satellite Download ────────────────────────────────────
sentinelsat==1.2.1
eodag==2.12.1

# ─── SAR Processing ────────────────────────────────────────
snapista==0.5.0
pysarpro==0.2.0

# ─── ML / CV ───────────────────────────────────────────────
ultralytics==8.1.47
torch==2.2.2
torchvision==0.17.2
opencv-python-headless==4.9.0.80
scikit-image==0.23.2
timm==0.9.16

# ─── Gemini AI ─────────────────────────────────────────────
google-generativeai==0.5.2

# ─── Browser Automation ────────────────────────────────────
playwright==1.43.0

# ─── PDF Parsing ───────────────────────────────────────────
pdfplumber==0.11.0
camelot-py[base]==0.11.0

# ─── Weather ───────────────────────────────────────────────
herbie-data==2024.3.0

# ─── Object Storage ────────────────────────────────────────
minio==7.2.5

# ─── Real-time ─────────────────────────────────────────────
python-socketio==5.11.2

# ─── Drone Telemetry ──────────────────────────────────────
pymavlink==2.4.41

# ─── Logging / Monitoring ──────────────────────────────────
structlog==24.1.0
prometheus-fastapi-instrumentator==7.0.0

# ─── MLflow ─────────────────────────────────────────────────
mlflow==2.12.1

# ─── HTTP ───────────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1
